When cloning esmf, use full depth
Using `--depth 1` is problematic when building the ESMPy documentation
because (unless the latest commit has been tagged) it means that no tags
are available from which the version can be determined. Cloning with
full depth only takes a few extra seconds, so do that.

It may be okay to use `--depth 1` for the other clones - for building
the ESMF documentation and for test-coverage - but it seems best to keep
these consistent, and since the cost is fairly low, this seems safest.